+++ flambe/compile/registry.py
@@ -1,12 +1,11 @@
 from typing import Type, TypeVar, Callable, Mapping, Dict, List, Any, Optional, Set, NamedTuple, Sequence, Iterable, Union
 from abc import abstractmethod, ABC
 from collections import defaultdict
-from warnings import warn
 import functools
 import logging
 import inspect
 
-from ruamel.yaml import YAML, ScalarNode
+from ruamel.yaml import YAML
 
 from flambe.compile.common import Singleton
 
@@ -15,7 +14,7 @@ ROOT_NAMESPACE = ''
 
 
 class RegistryEntry(NamedTuple):
-    """Entry in the registry representing a class, tags, factories"""
+    """Entry in the registry representing a callable, tags, factories"""
 
     callable: Callable
     default_tag: str
@@ -33,14 +32,14 @@ class RegistrationError(Exception):
 
 # Maps from class to registry entry, which contains the tags, aliases,
 # and factory methods
-SubRegistry = Dict[Type, RegistryEntry]
+SubRegistry = Dict[Callable, RegistryEntry]
 
 
 class Registry(metaclass=Singleton):
 
     def __init__(self):
         self.namespaces: Dict[str, SubRegistry] = defaultdict(SubRegistry)
-        self.callable_to_namespaces: Dict[type, Sequence[str]] = defaultdict(list)
+        self.callable_to_namespaces: Dict[Callable, Sequence[str]] = defaultdict(list)
 
     def create(self,
                callable: Callable,
@@ -56,8 +55,11 @@ class Registry(metaclass=Singleton):
         if tags is not None and isinstance(tags, list) and len(tags) < 1:
             raise ValueError('At least one tag must be specified. If the default (class name) '
                              'desired, pass in nothing or None.')
-        if factories is not None and len(factories) < 1:
-            raise ValueError('At least one factory must be specified if any are given.')
+        if factories is not None:
+            if len(factories) < 1:
+                raise ValueError('At least one factory must be specified if any are given.')
+            if not isinstance(callable, type):
+                raise ValueError('Callable must be a class if factories are specified')
         try:
             tags = tags or [callable.__name__]
         except AttributeError:
@@ -132,8 +134,20 @@ class Registry(metaclass=Singleton):
         for class_, namespace in self.class_to_namespace.items():
             yield self.namespaces[namespace][class_]
 
+    def pretty_str(self) -> str:
+        output = ""
+        for reg_entry in self:
+            pass
+
 
 def get_registry():
+    """Future proof entry point for registry access
+
+    Eventually we might want to refactor again or even sync registry
+    across machines, support multi-threading etc. so that logic can
+    appear here.
+
+    """
     return Registry()
 
 
@@ -150,7 +164,7 @@ def get_class_namespace(class_: Type):
         return ROOT_NAMESPACE
 
 
-def register(cls: Type[A],
+def register_class(cls: Type[A],
              tag: str,
              from_yaml: Optional[Callable] = None,
              to_yaml: Optional[Callable] = None) -> Type[A]:
@@ -178,24 +192,6 @@ def alias(tag: str) -> Callable[[RT], RT]:
         return cls
 
     return decorator
-
-
-class Registrable:
-    """Subclasses automatically registered as yaml tags
-
-    Automatically registers subclasses with the yaml loader by
-    adding a constructor and representer which can be overridden
-    """
-
-    def __init_subclass__(cls: Type['Registrable'],
-                          should_register: Optional[bool] = True,
-                          tag_override: Optional[str] = None,
-                          from_yaml: Optional[Callable] = None,
-                          to_yaml: Optional[Callable] = None,
-                          **kwargs: Mapping[str, Any]) -> None:
-        super().__init_subclass__(**kwargs)  # type: ignore
-        if should_register:
-            register(cls, tag_override, from_yaml, to_yaml)
 
 
 class registrable_factory:
@@ -234,106 +230,3 @@ class registrable_factory:
         except:
             pass
         setattr(owner, name, self.fn)
-
-
-class MappedRegistrable(Registrable):
-
-    @classmethod
-    def to_yaml(cls, representer: Any, node: Any, tag: str) -> Any:
-        """Use representer to create yaml representation of node"""
-        return representer.represent_mapping(tag, node._saved_kwargs)
-
-    @classmethod
-    def from_yaml(cls, constructor: Any, node: Any, factory_name: str) -> Any:
-        """Use constructor to create an instance of cls"""
-        if inspect.isabstract(cls):
-            msg = f"You're trying to initialize an abstract class {cls.__name__}. " \
-                  + "If you think it's concrete, double check you've spelled " \
-                  + "all the originally abstract method names correctly."
-            raise Exception(msg)
-        if isinstance(node, ScalarNode):
-            nothing = constructor.construct_yaml_null(node)
-            if nothing is not None:
-                warn(f"Non-null scalar argument to {cls.__name__} will be ignored. A map of kwargs"
-                     " should be used instead.")
-            return cls()
-        # NOTE: construct_yaml_map is a generator that yields the
-        # constructed data and then updates it
-        kwargs, = list(constructor.construct_yaml_map(node))
-        if factory_name is not None:
-            factory_method = getattr(cls, factory_name)
-        else:
-            factory_method = cls
-        instance = factory_method(**kwargs)
-        instance._saved_kwargs = kwargs
-        return instance
-
-
-class Schema(MutableMapping[str, Any]):
-
-    def __init__(self, flambe_callable: Callable, flambe_factory_name: Optional[str] = None, **kwargs: Any):
-        self.callable = callable
-        if flambe_factory_name is None:
-            if not isinstance(self.callable, type):
-                raise NotImplementedError('Using non-class callables with Schema is not yet supported')
-            self.factory_method = flambe_callable
-        else:
-            if not isinstance(self.callable, type):
-                raise Exception(f'Cannot specify factory name on non-class callable {callable}')
-            self.factory_method = getattr(self.callable, flambe_factory_name)
-        self.kwargs = kwargs
-        self.output_object = None
-        self.created_with_tag = None
-
-    @classmethod
-    def from_yaml(cls, callable: Callable, constructor: Any, node: Any, factory_name: str) -> Any:
-        """Use constructor to create an instance of cls"""
-        pass
-
-    @classmethod
-    def to_yaml(cls, representer: Any, node: Any, tag: str) -> Any:
-        """Use representer to create yaml representation of node"""
-        pass
-
-    def __call__(self, root):
-        if self.output_object is not None:
-            return self.output_object
-        kwargs = fill_defaults(self.factory, self.kwargs)
-        self.kwargs.update(kwargs)
-
-        def helper(obj):
-            if isinstance(obj, Link):
-                return obj(root)
-            elif isinstance(obj, Schema):
-                return obj(root)
-            elif isinstance(obj, dict):
-                return {k: helper(v) for k, v in obj.items()}
-            elif isinstance(obj, list):
-                [helper(x) for x in obj]
-            else:
-                return obj
-
-        kwargs = helper(kwargs)
-
-        if isinstance(self.callable, Component):
-            # TODO run hooks
-            pass
-        output = self.factory_method(**kwargs)
-        self.output_object = output
-        return output
-
-
-def add_callable_from_yaml(from_yaml_fn: Callable, callable: Callable) -> Callable:
-    @functools.wraps(from_yaml_fn)
-    def wrapped(constructor: Any, node: Any) -> Any:
-        obj = from_yaml_fn(constructor, node, callable=callable)
-        return obj
-    return wrapped
-
-
-class Schematic(Registrable):
-
-    def __init_subclass__(cls: Type['Registrable'],
-                          **kwargs: Mapping[str, Any]) -> None:
-        from_yaml_fn = add_callable_from_yaml(Schema.from_yaml, callable=cls)
-        super().__init_subclass__(from_yaml=Schema.from_yaml, to_yaml=Schema.to_yaml, **kwargs)
